+++ src/LSST/M1M3/TS/Commands/Update.cpp
@@ -31,6 +31,9 @@
 #include "Events/Heartbeat.h"
 #include "Events/SummaryState.h"
 #include "Events/ThermalInfo.h"
+#include "Events/ThermalWarning.h"
+
+#include "Settings/Heaters.h"
 
 #include "Telemetry/MixingValve.h"
 #include "Telemetry/ThermalData.h"
@@ -50,6 +53,7 @@ LSST::cRIO::task_return_t Update::run() {
     _sendMixingValve();
 
     Events::EnabledILC::instance().send();
+    Events::ThermalWarning::instance().send();
 
     SPDLOG_TRACE("Commands::Update leaving execute");
 
@@ -79,7 +83,7 @@ void Update::_sendMixingValve() {
 
 void Update::_sendFCU() {
     /// State of the state machine handling bus recovery from power failure
-    static enum { OK, FAILED, RESET_ERROR, STANDBY, SERVER_ID, DISABLED, ENABLED, SET_FCUS } _bus_state = OK;
+    static enum { OK, FAILED, RESET_ERROR, STANDBY, SERVER_ID, DISABLED, ENABLED } _bus_state = OK;
 
     static auto next_update = std::chrono::steady_clock::now() - 20ms;
 
@@ -102,13 +106,11 @@ void Update::_sendFCU() {
 
         switch (_bus_state) {
             case OK:
-                app.callFunctionOnAllIlcs([&app](uint8_t address) {
-                    if (Events::SummaryState::instance().active()) {
-                        app.ilc()->reportThermalStatus(address);
-                    } else {
-                        app.ilc()->reportServerStatus(address);
-                    }
-                });
+                if (Events::SummaryState::instance().active()) {
+                    app.callFunctionOnAllIlcs(
+                            [&app](int address) { app.ilc()->reportThermalStatus(address); });
+                }
+                app.callFunctionOnAllIlcs([&app](int address) { app.ilc()->reportServerStatus(address); });
                 break;
             case FAILED:
                 app.callFunctionOnAllIlcs([&app](uint8_t address) {
@@ -134,10 +136,6 @@ void Update::_sendFCU() {
                 app.callFunctionOnAllIlcs(
                         [&app](uint8_t address) { app.ilc()->reportServerStatus(address); });
                 break;
-            case SET_FCUS:
-                Events::FcuTargets::instance().recover();
-                app.ilc()->clear();
-                break;
             default:
                 SPDLOG_ERROR("Reached invalid ILC bus state: {}", static_cast<int>(_bus_state));
                 Events::SummaryState::set_state(MTM1M3TS::MTM1M3TS_shared_SummaryStates_FaultState);
@@ -156,6 +154,7 @@ void Update::_sendFCU() {
                 break;
             case FAILED:
                 _bus_state = RESET_ERROR;
+                Settings::Heaters::instance().reset_FCU_PIDs();
                 break;
             case RESET_ERROR:
                 _bus_state = STANDBY;
@@ -171,9 +170,6 @@ void Update::_sendFCU() {
                 _bus_state = ENABLED;
                 break;
             case ENABLED:
-                _bus_state = SET_FCUS;
-                break;
-            case SET_FCUS:
                 _bus_state = OK;
                 break;
         }